refactor: split footsteps layer modules
