+++ humans-globe/components/footsteps/FootstepsViz.tsx
@@ -3,7 +3,12 @@
 import { useState, useMemo, useEffect, memo, useCallback, useRef } from 'react';
 import { getViewMode, setViewMode } from '@/lib/viewModeStore';
 import { getLODLevel } from '@/lib/lod';
-import { createHumanTilesLayer, createStaticTerrainLayer, createPlainBackgroundLayers, radiusStrategies } from '@/components/footsteps/layers/layers';
+import {
+  createHumanTilesLayer,
+  createStaticTerrainLayer,
+  createPlainBackgroundLayers,
+  radiusStrategies,
+} from '@/components/footsteps/layers';
 import { type LayersList } from '@deck.gl/core';
 import SupportingText from '@/components/footsteps/overlays/SupportingText';
 import LegendOverlay from '@/components/footsteps/overlays/LegendOverlay';
@@ -14,7 +19,10 @@ import useGlobeViewState from '@/components/footsteps/hooks/useGlobeViewState';
 import VizToggles from '@/components/ui/VizToggles';
 
 type PickingInfo = {
-  object?: { properties?: { population?: number }; geometry?: { coordinates?: [number, number] } };
+  object?: {
+    properties?: { population?: number };
+    geometry?: { coordinates?: [number, number] };
+  };
   x?: number;
   y?: number;
 };
@@ -26,13 +34,14 @@ interface FootstepsVizProps {
 function FootstepsViz({ year }: FootstepsVizProps) {
   // View mode toggle state with cookie persistence for SSR compatibility
   const [is3DMode, setIs3DMode] = useState(() => getViewMode());
-  
+
   // Terrain toggle state - default to plain mode for better dot visibility
   const [showTerrain, setShowTerrain] = useState(false);
-  
+
   // Simplified viewState management - single state for both modes
-  const { viewState, onViewStateChange, isZooming, isPanning } = useGlobeViewState();
-  
+  const { viewState, onViewStateChange, isZooming, isPanning } =
+    useGlobeViewState();
+
   // Population tooltip state
   const [tooltipData, setTooltipData] = useState<{
     population: number;
@@ -41,14 +50,11 @@ function FootstepsViz({ year }: FootstepsVizProps) {
     settlementType?: string;
     clickPosition: { x: number; y: number };
   } | null>(null);
-  
-  
+
   // Save view mode preference to cookie
   useEffect(() => {
     setViewMode(is3DMode);
   }, [is3DMode]);
-  
-  
 
   // Removed complex viewport bounds system - tiles handle spatial filtering efficiently
 
@@ -68,7 +74,6 @@ function FootstepsViz({ year }: FootstepsVizProps) {
   const newLayerHasTileRef = useRef<boolean>(false);
   const YEAR_FADE_MS = 300;
   const NEW_YEAR_FADE_MS = 120; // brief fade-in for perceptibility
-
 
   useEffect(() => {
     // If the year changed, set up a crossfade from the previous year to the new year.
@@ -102,35 +107,42 @@ function FootstepsViz({ year }: FootstepsVizProps) {
     setFeatureCount(0);
     setTotalPopulation(0);
   }, [year]);
-  
+
   // Pre-render guard: if the year just changed on this render,
   // keep the previous year visible and hide the new year until tiles arrive.
   // This prevents a blank frame where the new-year layer (opacity 1) has no tiles yet
   // and the previous-year layer has not been added by the effect.
   const yearChangedThisRender = year !== prevPropYearRef.current;
-  const renderPrevYear = yearChangedThisRender ? prevPropYearRef.current : prevYear;
+  const renderPrevYear = yearChangedThisRender
+    ? prevPropYearRef.current
+    : prevYear;
   const renderCurrentOpacity = yearChangedThisRender ? 0 : currentYearOpacity;
   const renderPrevOpacity = yearChangedThisRender ? 1 : prevYearOpacity;
 
   // Background layers - terrain or plain based on toggle
   const backgroundLayers = useMemo(() => {
-    return showTerrain ? [createStaticTerrainLayer()] : createPlainBackgroundLayers();
+    return showTerrain
+      ? [createStaticTerrainLayer()]
+      : createPlainBackgroundLayers();
   }, [showTerrain]);
-  
-  
+
   // Stable LOD level for memoization - only changes at discrete boundaries
   const roundedZoom = Math.floor(viewState.zoom);
   const stableLODLevel = useMemo(() => {
     return getLODLevel(roundedZoom);
   }, [roundedZoom]);
-  
+
   // Simplified: use viewState directly since LOD system provides stability
   const layerViewState = viewState;
-  
+
   // Hoisted helper so it's available to callbacks declared below
-  function featuresFromTile(tile: any): Array<{ properties?: { population?: number } }> {
+  function featuresFromTile(
+    tile: any,
+  ): Array<{ properties?: { population?: number } }> {
     try {
-      const tileCoords = tile?.index ? `${tile.index.z}/${tile.index.x}/${tile.index.y}` : 'unknown';
+      const tileCoords = tile?.index
+        ? `${tile.index.z}/${tile.index.x}/${tile.index.y}`
+        : 'unknown';
       void tileCoords; // quiet unused in production
       const possibleFeatures = [
         tile?.data?.features,
@@ -156,125 +168,172 @@ function FootstepsViz({ year }: FootstepsVizProps) {
     }
   }
 
-  const createHumanLayerForYear = useCallback((targetYear: number, lodLevel: number, layerOpacity: number, instanceId: string, isNewYearLayer: boolean) => {
-    const radiusStrategy = is3DMode ? radiusStrategies.globe3D : radiusStrategies.zoomAdaptive;
+  const createHumanLayerForYear = useCallback(
+    (
+      targetYear: number,
+      lodLevel: number,
+      layerOpacity: number,
+      instanceId: string,
+      isNewYearLayer: boolean,
+    ) => {
+      const radiusStrategy = is3DMode
+        ? radiusStrategies.globe3D
+        : radiusStrategies.zoomAdaptive;
 
-    // Crossfade policy:
-    // - Before new tiles are ready: no transitions (avoid enter fade on prev layer)
-    // - When crossfade starts: NEW fades in briefly (NEW_YEAR_FADE_MS) for visibility,
-    //   PREV fades out (YEAR_FADE_MS)
-    const fadeMs = newLayerReadyRef.current
-      ? (isNewYearLayer ? NEW_YEAR_FADE_MS : YEAR_FADE_MS)
-      : 0;
+      // Crossfade policy:
+      // - Before new tiles are ready: no transitions (avoid enter fade on prev layer)
+      // - When crossfade starts: NEW fades in briefly (NEW_YEAR_FADE_MS) for visibility,
+      //   PREV fades out (YEAR_FADE_MS)
+      const fadeMs = newLayerReadyRef.current
+        ? isNewYearLayer
+          ? NEW_YEAR_FADE_MS
+          : YEAR_FADE_MS
+        : 0;
 
-    return createHumanTilesLayer(
-      targetYear,
-      lodLevel,
-      layerViewState,
-      radiusStrategy,
-      // onClick
-      (raw: unknown) => {
-        const info = raw as PickingInfo;
-        if (info?.object) {
-          const f = info.object as any;
-          const population = f?.properties?.population || 0;
-          const coordinates = (f?.geometry?.coordinates as [number, number]) || [0, 0];
-          const clickPosition = { x: info.x || 0, y: info.y || 0 };
-          setTooltipData({ population, coordinates, year: targetYear, clickPosition });
-        }
-      },
-      // onHover
-      (raw: unknown) => {
-        const info = raw as PickingInfo;
-        if (info?.object) {
-          const f = info.object as any;
-          const population = f?.properties?.population || 0;
-          const coordinates = (f?.geometry?.coordinates as [number, number]) || [0, 0];
-          const clickPosition = { x: info.x || 0, y: info.y || 0 };
-          setTooltipData({ population, coordinates, year: targetYear, clickPosition });
-        } else {
-          setTooltipData(null);
-        }
-      },
-      {
-        onTileLoad: (_tile: unknown) => {
-          // Only consider tile readiness/loading state from the NEW year layer.
-          // Do NOT start crossfade here; wait for onViewportLoad to ensure full coverage.
-          if (isNewYearLayer) {
-            setTileLoading(false);
-            if (!newLayerHasTileRef.current) {
-              newLayerHasTileRef.current = true;
-            }
+      return createHumanTilesLayer(
+        targetYear,
+        lodLevel,
+        layerViewState,
+        radiusStrategy,
+        // onClick
+        (raw: unknown) => {
+          const info = raw as PickingInfo;
+          if (info?.object) {
+            const f = info.object as any;
+            const population = f?.properties?.population || 0;
+            const coordinates = (f?.geometry?.coordinates as [
+              number,
+              number,
+            ]) || [0, 0];
+            const clickPosition = { x: info.x || 0, y: info.y || 0 };
+            setTooltipData({
+              population,
+              coordinates,
+              year: targetYear,
+              clickPosition,
+            });
           }
         },
-        onViewportLoad: (rawTiles: unknown[]) => {
-          try {
-            // Update metrics from the new (target) year layer when it reports tiles
+        // onHover
+        (raw: unknown) => {
+          const info = raw as PickingInfo;
+          if (info?.object) {
+            const f = info.object as any;
+            const population = f?.properties?.population || 0;
+            const coordinates = (f?.geometry?.coordinates as [
+              number,
+              number,
+            ]) || [0, 0];
+            const clickPosition = { x: info.x || 0, y: info.y || 0 };
+            setTooltipData({
+              population,
+              coordinates,
+              year: targetYear,
+              clickPosition,
+            });
+          } else {
+            setTooltipData(null);
+          }
+        },
+        {
+          onTileLoad: (_tile: unknown) => {
+            // Only consider tile readiness/loading state from the NEW year layer.
+            // Do NOT start crossfade here; wait for onViewportLoad to ensure full coverage.
             if (isNewYearLayer) {
-              const tiles = rawTiles as Array<unknown>;
-              let count = 0;
-              let pop = 0;
-              for (const t of tiles) {
-                // eslint-disable-next-line @typescript-eslint/no-explicit-any
-                const tile = t as any;
-                const feats = featuresFromTile(tile);
-                count += feats.length;
-                for (const g of feats) pop += Number(g?.properties?.population) || 0;
-              }
-              setFeatureCount(count);
-              setTotalPopulation(pop);
               setTileLoading(false);
-
-              // Start crossfade once new layer has some tiles
-              // Guard against race where onViewportLoad fires before isYearCrossfading updates
-              if (!newLayerReadyRef.current && (isYearCrossfading || yearChangedThisRender)) {
-                newLayerReadyRef.current = true;
-                setCurrentYearOpacity(1);
-                setPrevYearOpacity(0);
-                // Remove previous layer after fade completes
-                yearFadeTimeoutRef.current = setTimeout(() => {
-                  setPrevYear(null);
-                  setIsYearCrossfading(false);
-                  yearFadeTimeoutRef.current = null;
-                }, YEAR_FADE_MS + 50);
+              if (!newLayerHasTileRef.current) {
+                newLayerHasTileRef.current = true;
               }
             }
-          } catch (error) {
-            console.error(`[VIEWPORT-LOAD-ERROR] Year: ${targetYear}, LOD: ${lodLevel}:`, error);
+          },
+          onViewportLoad: (rawTiles: unknown[]) => {
+            try {
+              // Update metrics from the new (target) year layer when it reports tiles
+              if (isNewYearLayer) {
+                const tiles = rawTiles as Array<unknown>;
+                let count = 0;
+                let pop = 0;
+                for (const t of tiles) {
+                  // eslint-disable-next-line @typescript-eslint/no-explicit-any
+                  const tile = t as any;
+                  const feats = featuresFromTile(tile);
+                  count += feats.length;
+                  for (const g of feats)
+                    pop += Number(g?.properties?.population) || 0;
+                }
+                setFeatureCount(count);
+                setTotalPopulation(pop);
+                setTileLoading(false);
+
+                // Start crossfade once new layer has some tiles
+                // Guard against race where onViewportLoad fires before isYearCrossfading updates
+                if (
+                  !newLayerReadyRef.current &&
+                  (isYearCrossfading || yearChangedThisRender)
+                ) {
+                  newLayerReadyRef.current = true;
+                  setCurrentYearOpacity(1);
+                  setPrevYearOpacity(0);
+                  // Remove previous layer after fade completes
+                  yearFadeTimeoutRef.current = setTimeout(() => {
+                    setPrevYear(null);
+                    setIsYearCrossfading(false);
+                    yearFadeTimeoutRef.current = null;
+                  }, YEAR_FADE_MS + 50);
+                }
+              }
+            } catch (error) {
+              console.error(
+                `[VIEWPORT-LOAD-ERROR] Year: ${targetYear}, LOD: ${lodLevel}:`,
+                error,
+              );
+              setTileLoading(false);
+            }
+          },
+          onTileError: (error: unknown) => {
+            // eslint-disable-next-line @typescript-eslint/no-explicit-any
+            const errorInfo = error as any;
+            const tileCoords = errorInfo?.tile?.index || errorInfo?.index || {};
+            const url = errorInfo?.tile?.url || errorInfo?.url || 'unknown';
+            const status =
+              errorInfo?.response?.status || errorInfo?.status || 'unknown';
+            console.error(`[TILE-ERROR] Failed to load tile:`, {
+              year: targetYear,
+              lodLevel,
+              coords: `${tileCoords.z}/${tileCoords.x}/${tileCoords.y}`,
+              url,
+              status,
+              error: errorInfo?.message || errorInfo?.error || errorInfo,
+            });
             setTileLoading(false);
-          }
+          },
+          tileOptions: {
+            fadeMs: fadeMs, // drives opacity transitions per crossfade policy
+            debounceTime: isZooming || isPanning ? 80 : 20,
+            useBinary: false,
+          },
+          // Dev-only tint to make crossfade layers visually distinct during debugging
+          debugTint:
+            process.env.NODE_ENV !== 'production' && isYearCrossfading
+              ? isNewYearLayer
+                ? [0, 30, 80]
+                : [80, 0, 0]
+              : undefined,
         },
-        onTileError: (error: unknown) => {
-          // eslint-disable-next-line @typescript-eslint/no-explicit-any
-          const errorInfo = error as any;
-          const tileCoords = errorInfo?.tile?.index || errorInfo?.index || {};
-          const url = errorInfo?.tile?.url || errorInfo?.url || 'unknown';
-          const status = errorInfo?.response?.status || errorInfo?.status || 'unknown';
-          console.error(`[TILE-ERROR] Failed to load tile:`, {
-            year: targetYear,
-            lodLevel,
-            coords: `${tileCoords.z}/${tileCoords.x}/${tileCoords.y}`,
-            url,
-            status,
-            error: errorInfo?.message || errorInfo?.error || errorInfo
-          });
-          setTileLoading(false);
-        },
-        tileOptions: {
-          fadeMs: fadeMs, // drives opacity transitions per crossfade policy
-          debounceTime: (isZooming || isPanning) ? 80 : 20,
-          useBinary: false
-        },
-        // Dev-only tint to make crossfade layers visually distinct during debugging
-        debugTint: (process.env.NODE_ENV !== 'production' && isYearCrossfading)
-          ? (isNewYearLayer ? [0, 30, 80] : [80, 0, 0])
-          : undefined
-      },
-      layerOpacity,
-      instanceId,
-      is3DMode
-    );
-  }, [layerViewState, is3DMode, isZooming, isPanning, isYearCrossfading, yearChangedThisRender]);
+        layerOpacity,
+        instanceId,
+        is3DMode,
+      );
+    },
+    [
+      layerViewState,
+      is3DMode,
+      isZooming,
+      isPanning,
+      isYearCrossfading,
+      yearChangedThisRender,
+    ],
+  );
 
   // Fallback: if a tile arrived before crossfade flag turned true, start crossfade now.
   useEffect(() => {
@@ -284,14 +343,29 @@ function FootstepsViz({ year }: FootstepsVizProps) {
   }, [isYearCrossfading, year]);
 
   // Create human tiles layers for current and (if crossfading) previous year
-  const currentYearLayer = createHumanLayerForYear(year, stableLODLevel, renderCurrentOpacity, `human-layer-${year}`, true);
-  const previousYearLayer = renderPrevYear !== null ? createHumanLayerForYear(renderPrevYear as number, stableLODLevel, renderPrevOpacity, `human-layer-${renderPrevYear}`, false) : null;
-  
+  const currentYearLayer = createHumanLayerForYear(
+    year,
+    stableLODLevel,
+    renderCurrentOpacity,
+    `human-layer-${year}`,
+    true,
+  );
+  const previousYearLayer =
+    renderPrevYear !== null
+      ? createHumanLayerForYear(
+          renderPrevYear as number,
+          stableLODLevel,
+          renderPrevOpacity,
+          `human-layer-${renderPrevYear}`,
+          false,
+        )
+      : null;
+
   // Layer ordering: background layers -> settlement points
   const layers: LayersList = previousYearLayer
     ? ([...backgroundLayers, previousYearLayer, currentYearLayer] as LayersList)
     : ([...backgroundLayers, currentYearLayer] as LayersList);
-  
+
   try {
     // eslint-disable-next-line @typescript-eslint/no-explicit-any
     const logLayer = (layer: any, tag: string, tagYear: number | null) => {
@@ -309,15 +383,16 @@ function FootstepsViz({ year }: FootstepsVizProps) {
         fadeMs: opacityTransition?.duration,
         is3DMode,
         isYearCrossfading,
-        newLayerReady: newLayerReadyRef.current
+        newLayerReady: newLayerReadyRef.current,
       });
     };
     logLayer(currentYearLayer, 'current', year);
-    if (previousYearLayer) logLayer(previousYearLayer, 'previous', renderPrevYear as number);
+    if (previousYearLayer)
+      logLayer(previousYearLayer, 'previous', renderPrevYear as number);
   } catch {
     // ignore logging errors in dev
   }
-  
+
   return (
     <div className="relative w-full h-full">
       {is3DMode ? (
@@ -333,7 +408,7 @@ function FootstepsViz({ year }: FootstepsVizProps) {
           layers={layers}
         />
       )}
-      
+
       {/* Data info overlay */}
       <SupportingText
         loading={tileLoading}
@@ -343,14 +418,19 @@ function FootstepsViz({ year }: FootstepsVizProps) {
         lodLevel={stableLODLevel}
         lodEnabled={false}
         toggleLOD={() => {}}
-        renderMetrics={{loadTime: 0, processTime: 0, renderTime: 0, lastUpdate: 0}}
+        renderMetrics={{
+          loadTime: 0,
+          processTime: 0,
+          renderTime: 0,
+          lastUpdate: 0,
+        }}
         cacheSize={1}
         progressiveRenderStatus={undefined}
         viewportBounds={null}
         is3DMode={is3DMode}
         year={year}
       />
-      
+
       {/* View Mode + Terrain toggles */}
       <div className="absolute top-4 right-4 z-10 flex flex-col gap-2">
         <VizToggles
@@ -363,9 +443,9 @@ function FootstepsViz({ year }: FootstepsVizProps) {
 
       {/* Legend */}
       <LegendOverlay />
-      
+
       {/* Population Tooltip */}
-      <PopulationTooltip 
+      <PopulationTooltip
         data={tooltipData}
         onClose={() => setTooltipData(null)}
       />
