+++ humans-globe/components/footsteps/layers/index.ts
@@ -1,0 +1,10 @@
+export { radiusStrategies } from './radiusStrategies';
+export type { RadiusStrategy } from './radiusStrategies';
+
+export {
+  createBasemapLayer,
+  createEarthSphereLayer,
+  createStaticTerrainLayer,
+  createPlainBackgroundLayers,
+} from './backgroundLayers';
+export { createHumanTilesLayer } from './humanLayer';
